fix: update MinisiteViewer tests after feature rename
- Rename MinisiteDisplayFeatureTest.php → MinisiteViewerFeatureTest.php
- Update namespace from MinisiteDisplay to MinisiteViewer
- Update all class references from MinisiteDisplayFeature to MinisiteViewerFeature
- Update method names and test assertions for new feature name
- Fix docblock assertions to check for correct feature name

All 131 MinisiteViewer tests now passing (100% success rate)
